requirements.txt
streamlit==1.28.2
google-generativeai==0.8.3
pandas==2.2.1
easyocr==1.7.1
opencv-python-headless==4.9.0.80
numpy==1.24.4
reportlab==4.0.6